:truck: Move LogIntermediatePredictions callback into callbacks_wandb
Moving the LogIntermediatePredictions callback class from callbacks.py into callbacks_wandb.py.
