+++ trainer.py
@@ -16,8 +16,10 @@ from lightning.pytorch.callbacks import (
 from lightning.pytorch.cli import ArgsType, LightningCLI
 from lightning.pytorch.plugins.io import AsyncCheckpointIO
 
-from src.callbacks import LogIntermediatePredictions  # noqa: F401
-from src.callbacks_wandb import LogMAEReconstruction  # noqa: F401
+from src.callbacks_wandb import (  # noqa: F401
+    LogIntermediatePredictions,
+    LogMAEReconstruction,
+)
 from src.datamodule import ClayDataModule, GeoTIFFDataPipeModule  # noqa: F401
 from src.model_clay import CLAYModule
 from src.model_vit import ViTLitModule  # noqa: F401
